# schemas/actions/partner-operations.openapi.yaml
openapi: 3.1.0
info:
  title: Partner Command Center Operations
  version: 1.0.0
  description: High-utility trusted partner operations for GPT Actions.
servers:
  - url: https://partner-command-center-rho.vercel.app
security:
  - BearerAuth: []
paths:
...
  /api/partner-links:
    post:
      operationId: createPartnerTrackingLink
      summary: Create a partner tracking/referral link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [partner_id, destination_url]
              properties:
                partner_id: { type: string }
                destination_url: { type: string, format: uri }
                tracking_url: { type: string, format: uri }
                source: { type: string }
                medium: { type: string }
                campaign: { type: string }
                utm_source: { type: string }
                utm_medium: { type: string }
                utm_campaign: { type: string }
      responses:
        '201': { description: Tracking link created }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PARTNER_COMMAND_API_KEY
...